put popup into an iframe, to isolate css
This was quite a pain in the ass, as the iframe no longer automatically
resizes based on it's content. This then required a pretty ugly
workaround: the iframe sends it's size to the content script, which in
turn sets the size of the iframe. However, due to the iframe not
reporting the right size immediately, I had to resort to a polling based
method via window.setInterval. Pretty ugly :(