Keys are now generated and stored
Currently when you hit reset, the keys are not generated again.
I think this is ok, because the id of the device that was reset no longer
exists. The new id then gets the keys of the old device.
If you truly want to reset also the keys, you have to clear local
storage.